sync with upstream

+++ packages/strapi-connector-bookshelf/lib/queries.js
@@ -4,12 +4,17 @@
  */
 
 const _ = require('lodash');
+const { omit } = require('lodash/fp');
 const pmap = require('p-map');
 const { convertRestQueryParams, buildQuery, escapeQuery } = require('strapi-utils');
 const { contentTypes: contentTypesUtils } = require('strapi-utils');
 const { singular } = require('pluralize');
+const { handleDatabaseError } = require('./utils/errors');
+
+const BATCH_SIZE = 1000;
 
 const { PUBLISHED_AT_ATTRIBUTE } = contentTypesUtils.constants;
+const pickCountFilters = omit(['sort', 'limit', 'start']);
 
 module.exports = function createQueryBuilder({ model, strapi }) {
   /* Utils */
@@ -48,6 +53,14 @@ module.exports = function createQueryBuilder({ model, strapi }) {
     return db.transaction(trx => fn(trx));
   };
 
+  const wrapErrors = fn => async (...args) => {
+    try {
+      return await fn(...args);
+    } catch (error) {
+      return handleDatabaseError(error);
+    }
+  };
+
   /**
    * Find one entry based on params
    */
@@ -76,12 +89,12 @@ module.exports = function createQueryBuilder({ model, strapi }) {
   /**
    * Count entries based on filters
    */
-  function count(params = {}) {
-    const filters = convertRestQueryParams(_.omit(params, ['_sort', '_limit', '_start']));
+  function count(params = {}, { transacting } = {}) {
+    const filters = pickCountFilters(convertRestQueryParams(params));
 
     return model
       .query(buildQuery({ model, filters }))
-      .count()
+      .count({ transacting })
       .then(Number);
   }
 
@@ -136,7 +149,7 @@ module.exports = function createQueryBuilder({ model, strapi }) {
         return model.updateRelations({ id: entry.id, values: relations }, { transacting: trx });
       }
 
-      return this.findOne(params, null, { transacting: trx });
+      return findOne(params, null, { transacting: trx });
     };
 
     return wrapTransaction(runUpdate, { transacting });
@@ -162,7 +175,10 @@ module.exports = function createQueryBuilder({ model, strapi }) {
     return wrapTransaction(runDelete, { transacting });
   }
 
-  async function deleteMany(params, { transacting } = {}) {
+  async function deleteMany(
+    params,
+    { transacting, returning = true, batchSize = BATCH_SIZE } = {}
+  ) {
     if (params[model.primaryKey]) {
       const entries = await find({ ...params, _limit: 1 }, null, { transacting });
       if (entries.length > 0) {
@@ -171,12 +187,30 @@ module.exports = function createQueryBuilder({ model, strapi }) {
       return null;
     }
 
-    const paramsWithDefaults = _.defaults(params, { _limit: -1 });
-    const entries = await find(paramsWithDefaults, null, { transacting });
-    return pmap(entries, entry => deleteOne(entry.id, { transacting }), {
-      concurrency: 100,
-      stopOnError: true,
-    });
+    if (returning) {
+      const paramsWithDefaults = _.defaults(params, { _limit: -1 });
+      const entries = await find(paramsWithDefaults, null, { transacting });
+      return pmap(entries, entry => deleteOne(entry.id, { transacting }), {
+        concurrency: 100,
+        stopOnError: true,
+      });
+    }
+
+    // returning false, we can optimize the function
+    const batchParams = _.assign({}, params, { _limit: batchSize, _sort: 'id:ASC' });
+    // eslint-disable-next-line no-constant-condition
+    while (true) {
+      const batch = await find(batchParams, null, { transacting });
+
+      await pmap(batch, entry => deleteOne(entry.id, { transacting }), {
+        concurrency: 100,
+        stopOnError: true,
+      });
+
+      if (batch.length < BATCH_SIZE) {
+        break;
+      }
+    }
   }
 
   function search(params, populate) {
@@ -190,7 +224,8 @@ module.exports = function createQueryBuilder({ model, strapi }) {
   }
 
   function countSearch(params) {
-    const filters = convertRestQueryParams(_.omit(params, ['_q', '_sort', '_limit', '_start']));
+    const countParams = omit(['_q'], params);
+    const filters = pickCountFilters(convertRestQueryParams(countParams));
 
     return model
       .query(qb => qb.where(buildSearchQuery({ model, params })))
@@ -425,7 +460,6 @@ module.exports = function createQueryBuilder({ model, strapi }) {
         }
       }
     }
-    return;
   }
 
   async function deleteDynamicZoneOldComponents(entry, values, { key, joinModel, transacting }) {
@@ -458,7 +492,7 @@ module.exports = function createQueryBuilder({ model, strapi }) {
         };
       });
 
-    // verify the provided ids are realted to this entity.
+    // verify the provided ids are related to this entity.
     idsToKeep.forEach(({ id, component }) => {
       if (!allIds.find(el => el.id === id && el.component.uid === component.uid)) {
         const err = new Error(
@@ -520,7 +554,7 @@ module.exports = function createQueryBuilder({ model, strapi }) {
       .fetchAll({ transacting })
       .map(el => el.get('component_id').toString());
 
-    // verify the provided ids are realted to this entity.
+    // verify the provided ids are related to this entity.
     idsToKeep.forEach(id => {
       if (!allIds.includes(id)) {
         const err = new Error(
@@ -664,8 +698,8 @@ module.exports = function createQueryBuilder({ model, strapi }) {
   return {
     findOne,
     find,
-    create,
-    update,
+    create: wrapErrors(create),
+    update: wrapErrors(update),
     delete: deleteMany,
     count,
     search,
@@ -688,12 +722,14 @@ const buildSearchQuery = ({ model, params }) => qb => {
 
   const searchColumns = Object.keys(model._attributes)
     .filter(attribute => !associations.includes(attribute))
-    .filter(attribute => stringTypes.includes(model._attributes[attribute].type));
+    .filter(attribute => stringTypes.includes(model._attributes[attribute].type))
+    .filter(attribute => model._attributes[attribute].searchable !== false);
 
   if (!_.isNaN(_.toNumber(query))) {
     const numberColumns = Object.keys(model._attributes)
       .filter(attribute => !associations.includes(attribute))
-      .filter(attribute => numberTypes.includes(model._attributes[attribute].type));
+      .filter(attribute => numberTypes.includes(model._attributes[attribute].type))
+      .filter(attribute => model._attributes[attribute].searchable !== false);
     searchColumns.push(...numberColumns);
   }
